requirements.txt
camelot-py==0.11.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
et_xmlfile==2.0.0
ghostscript==0.7
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
pandas==2.2.3
pdfminer.six==20231228
pillow==11.0.0
pycparser==2.22
pypdf==5.1.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tabulate==0.9.0
tzdata==2024.2
